pyproject.toml
[project]
name = "wows-minimap-renderer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "wows-replay-parser",
    "pycairo>=1.26",
    "discord.py>=2.3",
    "click>=8.0",
    "rich>=13.0",
]

[project.scripts]
wows-render = "renderer.cli:main"

[tool.uv.sources]
wows-replay-parser = { path = "../wows-replay-parser", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["renderer", "bot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
